requirements.txt
py-cord
flask
python-dotenv